+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - Transformers
   - `iter.flat()` (like `Array.prototype.flat`)
+  - `iter.completeFlat()` for flattening an iterable completely.
 
 ## [2.3.0] - 2021-02-03